requirements.txt
streamlit==1.26.0
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0

numpy==1.26.1
matplotlib==3.8.0
plotly==5.18.0

certifi==2025.9.0
urllib3==2.0.5
